(wip) Make CPU 64-bit

+++ src/d2.rs
@@ -1,9 +1,9 @@
 use crate::intcode::IntCodeCPU;
 use std::collections::HashMap;
 
-const TARGET_PART2: i32 = 19690720;
+const TARGET_PART2: i64 = 19690720;
 
-pub fn part1() -> std::io::Result<i32> {
+pub fn part1() -> std::io::Result<i64> {
     let mut cpu = IntCodeCPU::new();
     cpu.read_data_file("input/d2.txt")?;
     cpu.set_data_at(1, 12);
@@ -12,7 +12,7 @@ pub fn part1() -> std::io::Result<i32> {
     return Ok(cpu.get_data_at(0));
 }
 
-pub fn part2() -> std::io::Result<i32> {
+pub fn part2() -> std::io::Result<i64> {
     // brute force
     for n in 0..100 {
         for v in 0..100 {
